requirements.txt
# Core dependencies
# tkinter
requests
beautifulsoup4
lxml
python-docx
reportlab
pandas
numpy
scikit-learn
sentence-transformers
faiss-cpu
chroma
openai
tiktoken

# Utility & system
python-dotenv
cryptography
loguru

# Testing
pytest
